+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   * Removed deprecated method `Client::getConnection()`. Use Client::getMongoClient().
   * Removed `Document::pushFromArray()`. Use `Document::pushEach()`.
   * Removed `Paginator::setQueryBuilder()’. Use `Paginator::setCursor()`.
+  * Removed `Document::fromArray()`. Use `Document::merge()` instead.
 
 ## 1.12.8 (2015-03-03)
   * Method 'Validator::validate()' marked as final